feat: Validate and convert last two digits of student ID
The code changes include validating and converting the last two digits of the student ID. If the last two digits contain non-numeric characters, an error message is displayed. Otherwise, the digits are converted to a list of integers. This improves the robustness of the code and ensures that only valid input is processed.